Move log panel to its own component directory.
This will make it easier to split the component in smaller pieces
