+++ src/client/components/logpanel/LogMessageComponent.vue
